validate for empty blocks: https://github.com/metanorma/metanorma-standoc/issues/1041
